dcloud-ai-agent/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.2
dashscope==1.20.11
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
duckduckgo_search==7.5.5
ecdsa==0.19.1
email_validator==2.2.0
Faker==37.0.2
fastapi>=0.68.0
fastapi-cli==0.0.7
flatbuffers==25.2.10
frozenlist==1.4.1
google_search_results==2.4.2
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.6.1
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-milvus==0.1.8
langchain-openai==0.3.11
langchain-redis==0.2.0
langchain-text-splitters==0.3.7
langsmith==0.3.19
lark==1.2.2
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
mdurl==0.1.2
milvus-lite==2.4.11
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.1
numpy==1.26.4
onnxruntime==1.19.2
openai==1.68.2
opencv-python==4.11.0.86
orjson==3.10.7
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
primp==0.14.0
propcache==0.2.0
protobuf==6.30.1
pyasn1==0.4.8
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.5.2
pydantic_core==2.18.4
Pygments==2.19.1
pymilvus==2.5.5
PyMySQL==1.1.0
pypdf==5.3.1
PyPDF2==3.0.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose[cryptography]>=3.1.0
python-multipart>=0.0.5
python-ulid==3.0.0
pytz==2025.1
PyYAML==6.0.2
rapidocr-onnxruntime==1.4.4
redis==5.2.1
redisvl==0.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
setuptools==76.1.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.46.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.66.5
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.2.3
uvicorn>=0.15.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
yarl==1.15.0
zstandard==0.23.0
